pyproject.toml
[tool.poetry]
name = "xparser"
version = "0.1.0"
description = ""
authors = ["Halone(Kirill) <randomnik192@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
peewee = "^3.16.0"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.6"
python-binance = "^1.0.17"
python-kucoin = "^2.1.3"
huobi_client = {git = "https://github.com/Halone228/huobi_Python"}
apscheduler = "^3.10.1"
peewee-async = "^0.8.0"
aiopg = "^1.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.7"}
loguru = "^0.6.0"
asyncpg = "^0.27.0"
bpython = "^0.24"
python-dotenv = "^1.0.0"
aiogram = "^2.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "xparser.main:run_dev"
test1 = "tests.test_1:run_tests"
